Use downloader instead of steamcmd-in-docker to build server image
Use depotdownloader instead of steamcmd to download the game data from
steam. This is slightly more reliable and runs on the host instead of
in a container. CS:S builds in 20 minutes instead of 30 and uses
slightly less space. In the future we can consider using
steam_depot_layer's manifest parameter to download a specific version
of the steam app.

- Set DisableAutoUpdate on SourceMod
- Add symlink for /root/.steam/sdk32/steamclient.so to avoid log spam
- Build base image with base.Dockerfile instead of rules_docker.